+++ ExtraTerminalCommands/CHANGELOG.md
@@ -1,11 +1,26 @@
+# V1.6.2
+
+## Additions:
+
+- You can now customize aliases to nearly all commands via configuration
+  - These aliases are client-side and won't be synced
+
+# V1.6.1
+
+## Bug Fixes:
+
+Prevent the aliased keywords from showing on the `other` commands menu.
+
 # V1.6.0
 
 ## Additions:
+
 - TP [player name] Command
 - Flash Command
 - Ping Command
 
 ## Removals:
+
 - A ton of aliases
   - Commands, Extra Commands, Extra-Commands, EC
   - R, R W
@@ -22,7 +37,9 @@
 # V1.5.0
 
 ## Additions:
+
 - Horn command (works with custom input)
 
 ## Bug fixes:
+
 - Fixed Extra command mismatch (in the extra menu it used to show "random [money], whilst money is not a filter)